se creo un propio componente para el navbar del perfil de usuario y algunas otras modificaciones al navbar general y ruta orders
