+++ client/src/views/Admin/Orders/Orders.jsx
@@ -1,0 +1,9 @@
+const Orders = () => {
+  return (
+    <div className="bg-red-300 flex-1 min-w-full">
+      Hola
+    </div>
+  )
+}
+
+export default Orders
